typos

+++ src/main/tests/test-triplet.cpp
@@ -50,8 +50,8 @@ namespace
 
 Y_UTEST(triplet)
 {
-    do_test<int>();
-    do_test<int64_t>();
+    do_test<int16_t>();
+    do_test<int32_t>();
     do_test<float>();
     do_test<double>();
